+++ src/AudioEngine.hpp
@@ -22,6 +22,12 @@ using string = std::string;
 
 namespace darksun {
 
+	struct SoundCategory {
+		string catName;
+		int volume;
+		float attentuationCharacteristic;
+	};
+
 	// Provides an audio engine interface
 	class AudioEngine {
 
@@ -74,6 +80,7 @@ namespace darksun {
 			string ref;
 			bool looped;
 			int startIndex;
+			int volume;
 		};
 
 		/* Private variables */
@@ -82,6 +89,8 @@ namespace darksun {
 
 		static std::vector<PlayRequest> soundsToPlay;
 		static std::vector<string> soundsToStop;
+
+		static std::map<string, SoundCategory> categories;
 
 	public:
 		// Init the engine
@@ -100,10 +109,39 @@ namespace darksun {
 		static void removeSound(string ref);
 
 		// Play the sound
-		static void playSound(string ref, bool loop = false, int startIndex = 0);
+		static void playSound(string ref, string cat, bool loop = false, int startIndex = 0);
 
 		// Stop the sound
 		static void stopSound(string ref);
+
+		// Add a new category
+		static void newCategory(string ref) {
+			SoundCategory cat;
+			cat.catName = ref;
+			cat.attentuationCharacteristic = 1.0f;
+			cat.volume = 100.0f;
+			categories[ref] = cat;
+		}
+
+		// Set the volume of a category
+		static void setCategoryVolume(string ref, float vol) {
+			if (vol < 0)
+				vol = 0;
+			if (vol > 100)
+				vol = 100;
+			if (categories.count(ref) == 0)
+				return;
+			categories[ref].volume = vol;
+		}
+
+		// Set the attenuation of a category
+		static void setCategoryAttenuation(string ref, float attenuation) {
+			if (attenuation < 0)
+				attenuation = 0;
+			if (categories.count(ref) == 0)
+				return;
+			categories[ref].attentuationCharacteristic = attenuation;
+		}
 	};
 
 }